Assignment 2
Moved my images to the header on each page. To make the images centered and set them up so if the size of the browser window is changed the images adapts. I created a div class for each image and set the margins/background color for each image in the layout css.