Assignment 5
I created a index that has a header, and a subheader. I added an article that has a few paragraphs. I designed the index for a mobile size device and set up a Media Queries so that when the size of the screen is larger than 769px the text will adapt to the size of the index.