Assignment11
I created a responsive image gallery that reacts to the size of the browser window. I set the images to be 15px apart and have some css hover effects for when the user moves over each image. I also added JavaScript for when the user clicks on each image the larger version of the image will appear. The user is also abale to click through the images without having to exit out of each larger image. I also created a button that uses JavaScript that will allow the user to switch from the winter themed page to the summer page.